requirements.txt
sentencepiece
fastapi
python-dotenv
nltk
torch
transformers
